# modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-site-setting-impl/rest-openapi.yaml
components:
...
    schemas:
...
        Error:
            properties:
                errorCode:
                    description:
                        Internal error code mapping
                    example: 996
                    readOnly: true
                    type: integer
                errorDescription:
                    example: Unable to find currency. Currency code should be expressed with
                        3-letter ISO 4217 format.
                    readOnly: true
                    type: string
                message:
                    example: No CommerceCurrency exists with the key {groupId=41811, code=US
                        Dollar}
                    readOnly: true
                    type: string
                status:
                    description:
                        HTTP Status code
                    example: 404
                    readOnly: true
                    type: integer
            required:
                - errorCode
                - errorDescription
                - message
                - status
            type: object
...
        Warehouse:
            properties:
                active:
                    example: true
                    type: boolean
                city:
                    example: Diamond Bar
                    type: string
                countryId:
                    example: 30130
                    format: int64
                    minimum: 0
                    type: integer
                description:
                    example: right stairs, first room on the left
                    type: string
                groupId:
                    example: 23130
                    format: int64
                    minimum: 0
                    readOnly: true
                    type: integer
                id:
                    example: 30130
                    format: int64
                    minimum: 0
                    type: integer
                latitude:
                    example: 33.9976884
                    format: double
                    type: number
                longitude:
                    example: -117.8144595
                    format: double
                    type: number
                mvccVersion:
                    example: 0
                    type: number
                name:
                    example: Warehouse Name
                    type: string
                primary:
                    example: true
                    type: boolean
                regionId:
                    example: 30234
                    format: int64
                    minimum: 0
                    type: integer
                street1:
                    example: 1400 Montefino Ave
                    type: string
                street2:
                    example: 1st floor
                    type: string
                street3:
                    example: suite 200
                    type: string
                zip:
                    example: "91765"
                    type: string
            required:
                - countryId
                - name
            type: object
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            scheme: bearer
            type: http
        OAuth2:
            flows:
                authorizationCode:
                    authorizationUrl: /oauth/authorize
                    scopes:
                        HeadlessCommerceAdminSiteSetting.admin: Grants access to admin operations
                        HeadlessCommerceAdminSiteSetting.read: Grants read access
                        HeadlessCommerceAdminSiteSetting.write: Grants write access
                    tokenUrl: /oauth/token
            type: oauth2
info:
    contact:
        email: team-commerce@liferay.com
        name: Commerce Team
    description:
        "Liferay Commerce Admin Site Setting API. A Java client JAR is available for use with the group ID
        'com.liferay', artifact ID 'com.liferay.headless.commerce.admin.site.setting.client', and version '4.0.0'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: Liferay Commerce Admin Site Setting API
    version: v1.0
openapi: 3.0.0
paths:
...
    /commerceAdminSiteSetting/{groupId}/warehouse:
        get:
            operationId: getCommerceAdminSiteSettingGroupWarehousePage
            parameters:
                - in: path
                  explode: false
                  name: groupId
                  required: true
                  schema:
                      description:
                          Group to filter (Liferay site)
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
                - in: query
                  explode: true
                  name: active
                  required: false
                  schema:
                      description:
                          Active to filter
                      example: true
                      type: boolean
                  style: form
                - in: query
                  explode: true
                  name: page
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  explode: true
                  name: pageSize
                  required: false
                  schema:
                      type: integer
                  style: form
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Warehouse"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Warehouse"
                                type: array
                    description:
                        Successful operation
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminSiteSetting.read
            summary: Gets a list of Warehouse.
            tags:
                - Warehouse
        post:
            operationId: postCommerceAdminSiteSettingGroupWarehouse
            parameters:
                - in: path
                  explode: false
                  name: groupId
                  required: true
                  schema:
                      description:
                          Group to filter (Liferay site)
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Warehouse"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Warehouse"
                required: true
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Warehouse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Warehouse"
                    description:
                        Created
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminSiteSetting.write
            summary: Creates a Warehouse.
            tags:
                - Warehouse
...
    /warehouse/{id}:
        delete:
            operationId: deleteWarehouse
            parameters:
                - in: path
                  explode: false
                  name: id
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            responses:
                204:
                    description:
                        No content
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminSiteSetting.write
            summary: Deletes a Warehouse by ID.
            tags:
                - Warehouse
        get:
            operationId: getWarehouse
            parameters:
                - in: path
                  explode: false
                  name: id
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Warehouse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Warehouse"
                    description:
                        Successful operation
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminSiteSetting.read
            summary: Gets a Warehouse by ID.
            tags:
                - Warehouse
        put:
            operationId: putWarehouse
            parameters:
                - in: path
                  explode: false
                  name: id
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Warehouse"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Warehouse"
                required: true
            responses:
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminSiteSetting.write
            summary: Updates a Warehouse by ID.
            tags:
                - Warehouse
servers:
    - description:
          Liferay Commerce API server
      url: https://commerce.{environment}.liferay.com/commerce-admin-site-setting/{basePath}
      variables:
          basePath:
              default: v1.0
          environment:
              default: api
              enum:
                  - api
                  - api.dev
                  - api.uat
    - description:
          Liferay Commerce Local Development (Hosts file)
      url: "{protocol}://{host}:{port}/o/commerce-admin-site-setting/{basePath}"
      variables:
          basePath:
              default: v1.0
          host:
              default: test.dev
          port:
              default: "8080"
              enum:
                  - "443"
                  - "8080"
                  - "8443"
          protocol:
              default: http
              enum:
                  - http
                  - https
    - description:
          Liferay Commerce Local Development with Ngrok
...